IMAD DONE

+++ ft_execution.c
@@ -6,7 +6,7 @@
 /*   By: idabligi <idabligi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/07 02:21:29 by idabligi          #+#    #+#             */
-/*   Updated: 2023/05/06 18:16:02 by idabligi         ###   ########.fr       */
+/*   Updated: 2023/05/06 18:44:25 by idabligi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,12 @@
 
 //---------------------------------------------------------------------------//
 
-int	ft_creatfile(void)
+int	ft_creatfile(t_list *data)
 {
+
+	if ((data->tatto == 5 || data->tatto == 6) && !ft_strcmp(data->next->arg, "/dev/stdout"))
+		return (1);
+
 	unlink("/tmp/input");
 	unlink("/tmp/output");
 	if (open("/tmp/input", O_CREAT, 0644) < 0)
@@ -46,7 +50,7 @@ void	ft_exec1(t_list *data, t_store *store, t_data *a)
 		{
 		store->path = ft_getpath(data->arg);
 		store->arg = ft_arg(data);
-        execve(store->path, store->arg, a->env22);
+		execve(store->path, store->arg, a->env22);
 		perror("execve");
 		exit(EXIT_FAILURE);
 		}
@@ -67,7 +71,7 @@ void	ft_exec1(t_list *data, t_store *store, t_data *a)
 
 void	ft_exec3(t_list *data, t_store *store, int i, int pid,t_data *a)
 {
-	if (!ft_creatfile())
+	if (!ft_creatfile(data))
 		return ;
 	while ((i <= store->count))
 	{
@@ -85,13 +89,13 @@ void	ft_exec3(t_list *data, t_store *store, int i, int pid,t_data *a)
 			while (data && (data->tatto != 4))
 			{
 				if ((data->tatto == 6) || (data->tatto == 8))
-					ft_creatfile();
+					ft_creatfile(data);
 				data = data->next;
 			}
 			if ((data) && (data->tatto == 4))
 				data = data->next;
 			if ((data) && (data->tatto == 5))
-				ft_creatfile();
+				ft_creatfile(data);
 			i++;
 		}
 	}
@@ -101,24 +105,24 @@ void	ft_exec3(t_list *data, t_store *store, int i, int pid,t_data *a)
 
 void	ft_execution(t_list *data, t_store *store,t_data *a , int fd)
 {
-    int save;
+	int save;
 
 	if (!(store->count))
 		return ;
 
-    if ((store->built) && !(store->pipe))
-    {
-        ft_checkinput(data, 0, 1, store);
-        if ((fd = ft_checkoutput(data, store, 1, 0)))
-        {
-            save = dup(STDOUT_FILENO);
-            dup2(fd, STDOUT_FILENO);
-            close(fd);
-        }
-        ft_execute_builtins(data, a);
-        dup2(save, STDOUT_FILENO);
-        close(save);
-    }
+	if ((store->built) && !(store->pipe))
+	{
+		ft_checkinput(data, 0, 1, store);
+		if ((fd = ft_checkoutput(data, store, 1, 0)))
+		{
+			save = dup(STDOUT_FILENO);
+			dup2(fd, STDOUT_FILENO);
+			close(fd);
+		}
+		ft_execute_builtins(data, a);
+		dup2(save, STDOUT_FILENO);
+		close(save);
+	}
 
 	else if (store->exec == 0)
 		ft_exec1(data, store,a);
